+++ shaders/libs/Water.frag
@@ -7,10 +7,10 @@
 const int ITER_GEOMETRY = 3;
 const int ITER_GEOMETRY2 = 4;
 
-float16_t sea_octave_micro(f16vec2 fuv, float16_t choppy) {
+float sea_octave_micro(vec2 fuv, float choppy) {
 	fuv += noise(fuv);
-	f16vec2 wv = 1.0-abs(sin(fuv));
-	f16vec2 swv = abs(cos(fuv));
+	vec2 wv = 1.0-abs(sin(fuv));
+	vec2 swv = abs(cos(fuv));
 	wv = mix(wv,swv,wv);
 	return pow(1.0-pow(wv.x * wv.y,0.75),choppy);
 }
@@ -18,35 +18,35 @@ float16_t sea_octave_micro(f16vec2 fuv, float16_t choppy) {
 const int ITER_GEOMETRY = 3;
 const int ITER_GEOMETRY2 = 3;
 
-float16_t sea_octave_micro(f16vec2 fuv, float16_t choppy) {
+float sea_octave_micro(vec2 fuv, float choppy) {
 	fuv += noise(fuv);
 	return (1.0 - sin(fuv.x)) * cos(1.0 - fuv.y) * 0.7;
 }
 #endif
-const float16_t SEA_CHOPPY = 4.5;
-const float16_t SEA_SPEED = 5.3;
-const float16_t SEA_FREQ = 0.11;
-const f16mat2 octave_m = f16mat2(1.4,1.1,-1.2,1.4);
+const float SEA_CHOPPY = 4.5;
+const float SEA_SPEED = 5.3;
+const float SEA_FREQ = 0.11;
+const mat2 octave_m = mat2(1.4,1.1,-1.2,1.4);
 
-const float16_t height_mul[4] = float[4] (
+const float height_mul[4] = float[4] (
 	0.32, 0.34, 0.30, 0.08
 );
-const float16_t total_height =
+const float total_height =
   height_mul[0] +
   height_mul[0] * height_mul[1] +
   height_mul[0] * height_mul[1] * height_mul[2] +
   height_mul[0] * height_mul[1] * height_mul[2] * height_mul[3] + 1.0;
-const float16_t rcp_total_height = 1.0 / total_height;
+const float rcp_total_height = 1.0 / total_height;
 
-float16_t getwave(vec3 p, in float16_t lod) {
-	float16_t freq = SEA_FREQ;
-	float16_t amp = SEA_HEIGHT;
-	float16_t choppy = SEA_CHOPPY;
-	f16vec2 fuv = p.xz * 2.0 - frameTimeCounter * vec2(0.1, 0.5); fuv.x *= 0.75;
+float getwave(vec3 p, in float lod) {
+	float freq = SEA_FREQ;
+	float amp = SEA_HEIGHT;
+	float choppy = SEA_CHOPPY;
+	vec2 fuv = p.xz * 2.0 - frameTimeCounter * vec2(0.1, 0.5); fuv.x *= 0.75;
 
-	float16_t wave_speed = frameTimeCounter * SEA_SPEED;
+	float wave_speed = frameTimeCounter * SEA_SPEED;
 
-	float16_t d, h = 0.0;
+	float d, h = 0.0;
 	for(int i = 0; i < ITER_GEOMETRY; i++) {
 		d = sea_octave_micro((fuv+wave_speed * vec2(0.1, 0.9))*freq,choppy);
 		h += d * amp;
@@ -57,15 +57,15 @@ float16_t getwave(vec3 p, in float16_t lod) {
 	return (h * rcp_total_height - SEA_HEIGHT) * lod;
 }
 
-float16_t getwave2(vec3 p, in float16_t lod) {
-	float16_t freq = SEA_FREQ;
-	float16_t amp = SEA_HEIGHT;
-	float16_t choppy = SEA_CHOPPY;
-	f16vec2 fuv = p.xz * 2.0 - frameTimeCounter * vec2(0.1, 0.5); fuv.x *= 0.75;
+float getwave2(vec3 p, in float lod) {
+	float freq = SEA_FREQ;
+	float amp = SEA_HEIGHT;
+	float choppy = SEA_CHOPPY;
+	vec2 fuv = p.xz * 2.0 - frameTimeCounter * vec2(0.1, 0.5); fuv.x *= 0.75;
 
-	float16_t wave_speed = frameTimeCounter * SEA_SPEED;
+	float wave_speed = frameTimeCounter * SEA_SPEED;
 
-	float16_t d, h = 0.0;
+	float d, h = 0.0;
 	for(int i = 0; i < ITER_GEOMETRY2; i++) {
 		d = sea_octave_micro((fuv+wave_speed * vec2(0.1, 0.9))*freq,choppy);
 		h += d * amp;
@@ -76,17 +76,17 @@ float16_t getwave2(vec3 p, in float16_t lod) {
 	return (h * rcp_total_height - SEA_HEIGHT) * lod;
 }
 
-f16vec3 get_water_normal(in f16vec3 wwpos, in float16_t displacement, in float16_t lod, in f16vec3 N, in f16vec3 T, in f16vec3 B) {
-	f16vec3 w1 = 0.01 * T + getwave2(wwpos + 0.01 * T, lod) * N;
-	f16vec3 w2 = 0.01 * B + getwave2(wwpos + 0.01 * B, lod) * N;
-	f16vec3 w0 = displacement * N;
+vec3 get_water_normal(in vec3 wwpos, in float displacement, in float lod, in vec3 N, in vec3 T, in vec3 B) {
+	vec3 w1 = 0.01 * T + getwave2(wwpos + 0.01 * T, lod) * N;
+	vec3 w2 = 0.01 * B + getwave2(wwpos + 0.01 * B, lod) * N;
+	vec3 w0 = displacement * N;
 	#define tangent w1 - w0
 	#define bitangent w2 - w0
 	return normalize(cross(bitangent, tangent));
 }
 
 #ifdef WATER_PARALLAX
-void WaterParallax(inout vec3 wpos, in float lod, in f16vec3 N) {
+void WaterParallax(inout vec3 wpos, in float lod, in vec3 N) {
 	const int maxLayers = 4;
 
 	wpos.y -= 1.62;
